Update test_jhu to handle asyncio
- Move test fixtures into tests/fixtures.py
- Update test fixtures to mock aiohttp.ClientSession.get, instead of requests.get
- Add a context manager to replace the global httputils.client_session with an AsyncMock
- Misc. cleanup
